banner_maker/requirements.txt
playwright==1.45.0
selectolax==0.3.21
openai>=1.80.0
Pillow==10.2.0
aiohttp==3.9.3
click==8.1.7
flask==3.0.0
werkzeug==3.0.1
gunicorn==21.2.0